% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-utils.R
\name{augment_aes}
\alias{augment_aes}
\title{Combine two aes() mappings}
\usage{
augment_aes(a1, a2)
}
\arguments{
\item{a1, a2}{the two aes mappings or lists}
}
\value{
combined aes mapping
}
\description{
Combine two aes() mappings
}
