% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_df.R
\name{calculate_bbox_polygon_df}
\alias{calculate_bbox_polygon_df}
\title{Calculate the bounding box of a polygon_df object}
\usage{
calculate_bbox_polygon_df(polygon_df)
}
\arguments{
\item{polygon_df}{polygon_df data.frame}
}
\value{
4-element numeric vector of \code{c(xmin, ymin, xmax, ymax)}
}
\description{
Calculate the bounding box of a polygon_df object
}
