% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-utils.R
\name{check_default}
\alias{check_default}
\title{Sanity check an argument value}
\usage{
check_default(
  x,
  options = NULL,
  default = NULL,
  type = NULL,
  prefix = "",
  verbose = FALSE
)
}
\arguments{
\item{x}{argument value}

\item{options}{valid options this value could take. If this is not NULL, then
x is checked to be a member of this set.  Default: NULL.}

\item{default}{the default value to use if the argument fails the checks.
If default is not NULL, then use it as the fallback value.
If default is NULL and options is not NULL, then use the first value
in \code{options}.}

\item{type}{NULL, 'num' or 'char'}

\item{prefix}{prefix of warning message.}

\item{verbose}{default: FALSE}
}
\value{
original \code{x} value if there are no issues, otherwise the \code{default}
        if given, otherwise the first element in \code{options}.
}
\description{
This is a milder, more explicit version of \code{match.arg} which gives a warning
and returns a good default.  match.arg() would throw errors which makes
geom/pattern development a bit trickier.
}
