% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick.R
\name{convert_img_to_array}
\alias{convert_img_to_array}
\title{Convert a magick image to an RGBA array.}
\usage{
convert_img_to_array(img)
}
\arguments{
\item{img}{magick image}
}
\value{
RGBA array with all values in range [0, 1]
}
\description{
This will promote gray or RGB images to RGBA arrays.
}
