% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_df.R
\name{convert_polygon_df_to_alpha_channel}
\alias{convert_polygon_df_to_alpha_channel}
\title{Convert a polygon to an alpha mask}
\usage{
convert_polygon_df_to_alpha_channel(polygon_df, width, height)
}
\arguments{
\item{polygon_df}{polygon_df object}

\item{width, height}{dimensions of the alpha channel matrix to return}
}
\value{
numeric matrix with all values with 0 or 1
}
\description{
Alpha values of 1 correspond to pixels within the polygon, all other
values will be zero.
}
