% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_df.R
\name{convert_polygon_df_to_polygon_grob}
\alias{convert_polygon_df_to_polygon_grob}
\title{Convert a \code{polygon_df} to \code{grid::polygonGrob} object}
\usage{
convert_polygon_df_to_polygon_grob(
  polygon_df,
  default.units = "npc",
  gp = gpar()
)
}
\arguments{
\item{polygon_df}{polygon_df data.frame}

\item{default.units}{'npc}

\item{gp}{default: gpar()}
}
\value{
sf polygon object
}
\description{
Convert a \code{polygon_df} to \code{grid::polygonGrob} object
}
