% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_df.R
\name{convert_polygon_df_to_polygon_sf}
\alias{convert_polygon_df_to_polygon_sf}
\title{Convert a \code{polygon_df} to an \code{sf} POLYGON/MULTIPOLYGON}
\usage{
convert_polygon_df_to_polygon_sf(polygon_df, buffer_dist = 0)
}
\arguments{
\item{polygon_df}{Polygon data.frame}

\item{buffer_dist}{buffer the polygon by the given distance}
}
\value{
sf polygon object
}
\description{
Convert a \code{polygon_df} to an \code{sf} POLYGON/MULTIPOLYGON
}
