% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_df.R
\name{convert_polygon_sf_to_polygon_df}
\alias{convert_polygon_sf_to_polygon_df}
\title{Convert a sf POLYGON/MULTIPOLYGON into a polygon_df}
\usage{
convert_polygon_sf_to_polygon_df(mp)
}
\arguments{
\item{mp}{sf POLYGON or MULTIPOLYGON object}
}
\value{
polygon_df data.frame object
}
\description{
Convert a sf POLYGON/MULTIPOLYGON into a polygon_df
}
