% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick.R
\name{convert_r_colour_to_magick_colour}
\alias{convert_r_colour_to_magick_colour}
\title{Convert an R colour to a magick colour}
\usage{
convert_r_colour_to_magick_colour(col)
}
\arguments{
\item{col}{may be a built in colour, like 'tomato' or a hex colour}
}
\value{
always returns a hex colour, except if col is NA when it retuns a special
magick colour 'none', which means transparent
}
\description{
Convert an R colour to a magick colour
}
