% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-circle.R
\name{create_circles_grob}
\alias{create_circles_grob}
\title{Create a cicleGrob object for a set of points}
\usage{
create_circles_grob(
  boundary_df,
  params,
  angle = 0,
  spacing = 0.1,
  density = 0.3,
  xoffset = 0,
  yoffset = 0,
  aspect_ratio
)
}
\arguments{
\item{boundary_df}{polygon_df data.frame}

\item{params}{params from the geom}

\item{angle}{angle of orientation (degrees)}

\item{spacing}{spacing in grid 'npc' coordinates. Usually in range [0, 1]}

\item{density}{fill fraction. Number in range [0, 1]}

\item{xoffset, yoffset}{offset the pattern creation origin.}

\item{aspect_ratio}{aspect_ratio}
}
\value{
A grid::circleGrob
}
\description{
Use 'sf' to help with the point in polygon intersections.
}
\details{
\itemize{
  \item{make grid to cover entire space}
  \item{rotate points into position}
  \item{create expanded boundary by r}
  \item{create contracted boundary by r}
  \item{remove all points outside the expanded boundary}
  \item{remove all points within contracted boundary -> internal circles}
  \item{any remaining points become part of the intersection grob}
  \item{total circles = treeGrob( internal_circls, intersection_circles)}
}
}
