% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-gradient.R
\name{create_gradient_as_array}
\alias{create_gradient_as_array}
\title{A shim to go between the main pattern function for an image, and the
specific pattern functon for an image.}
\usage{
create_gradient_as_array(width, height, params, legend)
}
\arguments{
\item{width, height}{image dimensions}

\item{params}{geom parameters from ggplot}

\item{legend}{logical. TRUE if call comes during legend creation, otherwise FALSE.}
}
\value{
RGBA array
}
\description{
This function really just unwraps the 'params' into better arguments and
passes them to the actual img/array creation function
}
