% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-gradient.R
\name{create_gradient_img}
\alias{create_gradient_img}
\title{create a gradient image as an array}
\usage{
create_gradient_img(
  width = 100,
  height = 100,
  colour1 = "red",
  colour2 = "blue",
  orientation = "vertical"
)
}
\arguments{
\item{width, height}{image dimensions}

\item{colour1, colour2}{gradient colours}

\item{orientation}{vertical, horizontal or radial}
}
\value{
magick image
}
\description{
create a gradient image as an array
}
