% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-.R
\name{create_key_pattern_grob}
\alias{create_key_pattern_grob}
\title{Create the patterned area to be used in the legend key}
\usage{
create_key_pattern_grob(data, params, size, aspect_ratio, boundary_df)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}

\item{aspect_ratio}{the geom's best guess at what the aspect ratio might be.}

\item{boundary_df}{the boundary of the pattern in npc coordinates}
}
\description{
Create the patterned area to be used in the legend key
}
