% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-magick-pattern.R
\name{create_magick_pattern_as_array}
\alias{create_magick_pattern_as_array}
\title{Read a user specified filename as an image}
\usage{
create_magick_pattern_as_array(width, height, params, legend)
}
\arguments{
\item{width}{image dimensions}

\item{height}{image dimensions}

\item{params}{geom parameters from ggplot}

\item{legend}{logical. TRUE if call comes during legend creation, otherwise FALSE.}
}
\value{
array
}
\description{
Read a user specified filename as an image
}
