% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-pattern.R
\name{create_magick_pattern_img}
\alias{create_magick_pattern_img}
\title{Create one of imagemagick's internal patterns}
\usage{
create_magick_pattern_img(
  width = 100,
  height = 100,
  type = "hexagons",
  colour = "black"
)
}
\arguments{
\item{width, height}{image dimensions}

\item{type}{name of the imagemagick pattern. See \url{http://www.imagemagick.org/script/formats.php}
for more information. See \link{magick_pattern_names} for a list of all
supported imagemagick patterns.}

\item{colour}{colour used to draw the pattern}
}
\description{
These are all 2-colour pixel patterns - the 'white' part will be made transparent.
}
