% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-pattern.R
\name{create_magick_pattern_img_scaled}
\alias{create_magick_pattern_img_scaled}
\title{Create a scaled version one of imagemagick's internal patterns}
\usage{
create_magick_pattern_img_scaled(
  width = 100,
  height = 100,
  type = "hexagons",
  colour = "black",
  scale = 1,
  filter = "box"
)
}
\arguments{
\item{width}{image dimensions}

\item{height}{image dimensions}

\item{type}{name of the imagemagick pattern. See \url{http://www.imagemagick.org/script/formats.php}
for more information. See \link{magick_pattern_names} for a list of all
supported imagemagick patterns.}

\item{colour}{colour used to draw the pattern}

\item{scale}{pattern scaling factor defualt: 1}

\item{filter}{filter to apply when sacling pattern.  default: 'box' for crisp
edges to the pixel.  See \code{magick::filter_types()} for a full list
of filters.}
}
\description{
These are all 2-colour pixel patterns - the 'white' part will be made transparent.
}
