% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-plasma.R
\name{create_magick_plasma_img}
\alias{create_magick_plasma_img}
\title{Create plasma using imagemagick}
\usage{
create_magick_plasma_img(width = 100, height = 100, colour)
}
\arguments{
\item{width, height}{image dimensions}

\item{colour}{colour}
}
\description{
Ref: \url{https://www.imagemagick.org/Usage/canvas/}
}
