% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array.R
\name{create_pattern_array}
\alias{create_pattern_array}
\title{Generate an RGBA array and create a raster grob to fill the region}
\usage{
create_pattern_array(params, boundary_df, aspect_ratio, legend, type = "image")
}
\arguments{
\item{params}{params/coords for a single element. named list or single row data.frame}

\item{boundary_df}{mask for the pattern rendering}

\item{aspect_ratio}{a aspect ratio of the plotting area.}

\item{legend}{is the pattern being created in the legend? default FALSE.
Use this flag if you want different pattern drawing behviour for the legend.}

\item{type}{type of image to fill with.  'image' for a file,
placeholder for an iamge from a web placeholder site,
gradient for an image magick gradient}
}
\value{
rasterGrob
}
\description{
the boundary of the shaded region will be used to generate a alpha mask
so that the array conforms to the shape.
}
