% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{create_pattern_grobs}
\alias{create_pattern_grobs}
\title{Generate all the pattern grobs for this geom}
\usage{
create_pattern_grobs(all_params, boundary_dfs, aspect_ratio)
}
\arguments{
\item{all_params}{parameters for all the elements}

\item{boundary_dfs}{boundary_df objects for each of the elements}

\item{aspect_ratio}{aspect ratio}
}
\value{
grobTree containing all the pattern grobs
}
\description{
Generate all the pattern grobs for this geom
}
