% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-circle.R, R/pattern-crosshatch-sf.R,
%   R/pattern-none.R, R/pattern-stripe-sf.R
\name{create_pattern_circles}
\alias{create_pattern_circles}
\alias{create_pattern_crosshatch_via_sf}
\alias{create_pattern_none}
\alias{create_pattern_stripes_via_sf}
\title{Create grob objects for the pattern elements within a boundary}
\usage{
create_pattern_circles(params, boundary_df, aspect_ratio, legend = FALSE)

create_pattern_crosshatch_via_sf(
  params,
  boundary_df,
  aspect_ratio,
  legend = FALSE
)

create_pattern_none(params, boundary_df, aspect_ratio, legend = FALSE)

create_pattern_stripes_via_sf(
  params,
  boundary_df,
  aspect_ratio,
  legend = FALSE
)
}
\arguments{
\item{params}{params/coords for a single element. named list or single row data.frame}

\item{boundary_df}{mask for the pattern rendering}

\item{aspect_ratio}{a aspect ratio of the plotting area.}

\item{legend}{is the pattern being created in the legend? default FALSE.
Use this flag if you want different pattern drawing behviour for the legend.}
}
\value{
grid grob objects.
}
\description{
Create grob objects for the pattern elements within a boundary
}
