% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-stripe-sf.R
\name{create_stripes_sf}
\alias{create_stripes_sf}
\title{Create a sf MULTIPOLYGON object where each polygon is an individual stripe.}
\usage{
create_stripes_sf(
  angle,
  spacing,
  density,
  xoffset = 0,
  yoffset = 0,
  aspect_ratio
)
}
\arguments{
\item{angle}{angle of orientation (degrees)}

\item{spacing}{spacing in grid 'npc' coordinates. Usually in range [0, 1]}

\item{density}{fill fraction. Number in range [0, 1]}

\item{xoffset}{offset the pattern creation origin.}

\item{yoffset}{offset the pattern creation origin.}

\item{aspect_ratio}{aspect_ratio}
}
\value{
`sf` multipolygon object
}
\description{
The stripes are created as polygons so that when clipped to rects/polygons,
the ends of the stripe are clipped correctly to the boundary.
}
