% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-placeholder.R
\name{fetch_placeholder_img}
\alias{fetch_placeholder_img}
\title{Fetch a placeholder image of the correct dimensions}
\usage{
fetch_placeholder_img(width = 100, height = 100, type = "kitten")
}
\arguments{
\item{width, height}{image dimensions}

\item{type}{specify the server from which to fetch images. default: 'kitten'}
}
\description{
Fetch a placeholder image from one of the sites used by web designers.  Append
'bw' to the name to fetch greyscale images instead of colour.
}
\details{
\describe{
\item{\code{kitten}}{ - Kittens from \url{https://placekitten.com/}}
\item{\code{picsum}}{ - Random images from \url{https://picsum.photos/}}
\item{\code{murray}}{ - Bill Murrary images from \url{http://www.fillmurray.com}}
\item{\code{cage}}{ - Nicholas Cage images from \url{https://www.placecage.com/}}
\item{\code{bear}}{ - Bears from \url{https://placebear.com/}}
\item{\code{seagal}}{ - Steven Seagal images from \url{https://www.stevensegallery.com/}}
\item{\code{flickr}}{ - Images from Flickr \url{https://loremflickr.com/}}
\item{\code{beard}}{ - Beards! \url{http://placebeard.it/}}
\item{\code{lorempixel}}{ - Random images from \url{http://lorempixel.com/}}
\item{\code{placeimg}}{ - Random images from \url{http://placeimg.com/}}
\item{\code{dummy}}{ - Numeric placeholder images from \url{https://dummyimage.com}}
}
}
