% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-fill.R
\name{fill_area_with_img_expand}
\alias{fill_area_with_img_expand}
\title{Resize image, preserving aspect, such that when cropped, the target width and height is achieved.}
\usage{
fill_area_with_img_expand(
  img,
  width,
  height,
  gravity = "Center",
  filter = "lanczos"
)
}
\arguments{
\item{img}{magick image}

\item{width}{target dimensions for final image}

\item{height}{target dimensions for final image}

\item{gravity}{how to anchor the image during the operation. See \code{magick::gravity_types()}}

\item{filter}{default: lanczos}
}
\value{
magick image of the required dimensions
}
\description{
Unless the width and height are carefully chosen, this operation will distort
the image to force it to fit the dimensions
}
\examples{
\dontrun{
filename <- system.file("img", "Rlogo.png", package="png")
img <- magick::image_read(filename)
fill_area_with_img_expand(img, 100, 400)
}
}
