% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-fill.R
\name{fill_area_with_img_fit}
\alias{fill_area_with_img_fit}
\title{Resize image, preserving aspect, such that one of target width or height is achieved without distoring or cropping.}
\usage{
fill_area_with_img_fit(
  img,
  width,
  height,
  gravity = "Center",
  filter = "lanczos"
)
}
\arguments{
\item{img}{magick image}

\item{width}{target dimensions for final image}

\item{height}{target dimensions for final image}

\item{gravity}{how to anchor the image during the operation. See \code{magick::gravity_types()}}

\item{filter}{default: lanczos}
}
\value{
magick image of the required dimensions
}
\description{
The canvas of the image will be expanded to cover the full width/height but
will be transparent.
}
\examples{
\dontrun{
filename <- system.file("img", "Rlogo.png", package="png")
img <- magick::image_read(filename)
fill_area_with_img_fit(img, 100, 400)
}
}
