% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-fill.R
\name{fill_area_with_img_tile}
\alias{fill_area_with_img_tile}
\title{Tile image to fill the specified area}
\usage{
fill_area_with_img_tile(img, width, height, filter = filter, scale = 1)
}
\arguments{
\item{img}{magick image}

\item{width}{target dimensions for final image}

\item{height}{target dimensions for final image}

\item{filter}{default: lanczos}

\item{scale}{scale the source image before tiling. default: 1}
}
\value{
magick image of the required dimensions
}
\description{
Unless the width and height are carefully chosen, this operation will distort
the image to force it to fit the dimensions
}
\examples{
\dontrun{
filename <- system.file("img", "Rlogo.png", package="png")
img <- magick::image_read(filename)
fill_area_with_img_none(img, 100, 400)
}
}
