% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-rect.R, R/geom-bar.R, R/geom-bin2d.R,
%   R/geom-boxplot.R, R/geom-col.R, R/geom-crossbar.R, R/geom-ribbon.R,
%   R/geom-density.R, R/geom-polygon.R, R/geom-map.R, R/geom-sf.R,
%   R/geom-tile.R, R/geom-violin.R
\name{geom_rect_pattern}
\alias{geom_rect_pattern}
\alias{geom_bar_pattern}
\alias{geom_bin2d_pattern}
\alias{geom_boxplot_pattern}
\alias{geom_col_pattern}
\alias{geom_crossbar_pattern}
\alias{geom_ribbon_pattern}
\alias{geom_area_pattern}
\alias{geom_density_pattern}
\alias{geom_polygon_pattern}
\alias{geom_map_pattern}
\alias{geom_sf_pattern}
\alias{geom_tile_pattern}
\alias{geom_violin_pattern}
\title{ggplot2 geoms with support for pattern fills}
\arguments{
\item{mapping, data, stat, position, linejoin, na.rm, show.legend, inherit.aes, binwidth}{See ggplot2 for description of the standard arguments to each geom.}
}
\description{
All geoms in this package are identical to their counterparts in ggplot2 except
that they can be filled with patterns.
}
\section{Pattern Arguments}{


Not all arguments apply to all patterns.

\describe{
\item{\strong{\code{pattern}}}{Pattern name string e.g. 'stripe' (default), 'crosshatch', 'point', 'circle', 'none'}
\item{\strong{\code{pattern_colour       }}}{ Colour used for strokes and points. default: 'black'}
\item{\strong{\code{pattern_fill         }}}{ Fill colour. default: 'black'}
\item{\strong{\code{pattern_angle        }}}{ Orientation of the pattern in degrees. default: 45}
\item{\strong{\code{pattern_density      }}}{ Approximate fill fraction of the pattern. Usually in range [0, 1], but can be higher. default: 0.2}
\item{\strong{\code{pattern_spacing      }}}{ Spacing of the pattern as a fraction of the plot size. default: 0.05}
\item{\strong{\code{pattern_xoffset,pattern_yoffset}}}{Offset the origin of the pattern. Range [0, 1]. default: 0.  Use this to slightly shift the origin of the pattern. For most patterns, the user should limit the offset value to be less than the pattern spacing.}
\item{\strong{\code{pattern_alpha        }}}{ Alpha transparency for pattern. default: 1}
\item{\strong{\code{pattern_linetype     }}}{ Stroke linetype. default: 1}
\item{\strong{\code{pattern_size         }}}{ Stroke line width. default: 1}
\item{\strong{\code{pattern_option_1     }}}{ Generic User value }
}
}

