% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{get_aspect_ratio_from_context}
\alias{get_aspect_ratio_from_context}
\title{Try and get an aspect ratio from the context of the plot}
\usage{
get_aspect_ratio_from_context(coord, panel_params)
}
\arguments{
\item{coord}{current "Coord"}

\item{panel_params}{panel params}
}
\value{
Aspect ratio
}
\description{
Try and get an aspect ratio from the context of the plot
}
