% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{get_aspect_ratio_from_coord}
\alias{get_aspect_ratio_from_coord}
\title{Get the aspect ratio from the coord}
\usage{
get_aspect_ratio_from_coord(coord, panel_params)
}
\arguments{
\item{coord}{ggplot2 Coord object}

\item{panel_params}{panel parameters}
}
\value{
aspect ratio or NULL if none present in the coord
}
\description{
Get the aspect ratio from the coord
}
