% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewport.R
\name{get_aspect_ratio_from_current_viewport}
\alias{get_aspect_ratio_from_current_viewport}
\title{Get the aspec ratio of  the current viewport}
\usage{
get_aspect_ratio_from_current_viewport()
}
\description{
Get the aspec ratio of  the current viewport
}
