% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpattern-defunct.R
\name{ggpattern-defunct}
\alias{ggpattern-defunct}
\alias{magick_filter_names}
\alias{magick_gravity_names}
\alias{magick_pattern_intensity_names}
\alias{magick_pattern_names}
\alias{magick_pattern_stripe_names}
\alias{placeholder_names}
\alias{calculate_bbox_polygon_df}
\alias{convert_img_to_array}
\alias{convert_polygon_df_to_alpha_channel}
\alias{convert_polygon_df_to_polygon_grob}
\alias{convert_polygon_df_to_polygon_sf}
\alias{convert_polygon_sf_to_polygon_df}
\alias{create_gradient_img}
\alias{fetch_placeholder_img}
\alias{fill_area_with_img}
\alias{rotate_polygon_df}
\title{Defunct data/functions}
\usage{
calculate_bbox_polygon_df(...)

convert_img_to_array(...)

convert_polygon_df_to_alpha_channel(...)

convert_polygon_df_to_polygon_grob(...)

convert_polygon_df_to_polygon_sf(...)

convert_polygon_sf_to_polygon_df(...)

create_gradient_img(...)

fetch_placeholder_img(...)

fill_area_with_img(...)

rotate_polygon_df(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
These data/functions are Defunct in this release of ggpattern.
\enumerate{
\item{For \code{magick_filter_names} use \code{magick::filter_types()} instead.}
\item{For \code{magick_gravity_names} use \code{magick::gravity_types()} instead.}
\item{For \code{magick_pattern_intensity_names} use \code{gridpattern::names_magick_intensity}.}
\item{For \code{magick_pattern_names} use \code{gridpattern::names_magick}.}
\item{For \code{magick_pattern_stripe_names} use \code{gridpattern::names_magick_stripe}.}
\item{For \code{placeholder_names} use \code{gridpattern::names_placeholder}.}
}
}
