% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpattern-deprecated.R
\name{ggpattern-deprecated}
\alias{ggpattern-deprecated}
\alias{magick_filter_names}
\alias{magick_gravity_names}
\alias{magick_pattern_intensity_names}
\alias{magick_pattern_names}
\alias{magick_pattern_stripe_names}
\alias{placeholder_names}
\title{Deprecated data/functions}
\description{
These data/functions are Deprecated in this release of ggpattern,
they will be marked as Defunct and removed in a future version.
}
\details{
\enumerate{
\item{For \code{magick_filter_names} use \code{magick::filter_types()} instead.}
\item{For \code{magick_gravity_names} use \code{magick::gravity_types()} instead.}
}
}
\examples{
    suppressWarnings({
      print(magick_filter_names)
      print(magick_gravity_names)
      print(magick_pattern_intensity_names)
      print(magick_pattern_names)
      print(magick_pattern_stripe_names)
      print(placeholder_names)
    })
}
