% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpattern-package.R
\docType{package}
\name{ggpattern}
\alias{ggpattern}
\title{ggpattern: A package of geoms with parameterised pattern fills}
\description{
ggpattern: A package of geoms with parameterised pattern fills
}
