% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-rect.R, R/geom-bar.R, R/geom-boxplot.R,
%   R/geom-col.R, R/geom-crossbar.R, R/geom-ribbon.R, R/geom-density.R,
%   R/geom-polygon.R, R/geom-map.R, R/geom-sf.R, R/geom-tile.R, R/geom-violin.R
\docType{data}
\name{GeomRectPattern}
\alias{GeomRectPattern}
\alias{GeomBarPattern}
\alias{GeomBoxplotPattern}
\alias{GeomColPattern}
\alias{GeomCrossbarPattern}
\alias{GeomRibbonPattern}
\alias{GeomAreaPattern}
\alias{GeomDensityPattern}
\alias{GeomPolygonPattern}
\alias{GeomMapPattern}
\alias{GeomSfPattern}
\alias{GeomTilePattern}
\alias{GeomViolinPattern}
\title{GeomRectPattern}
\description{
GeomRectPattern

GeomBarPattern

GeomBoxplotPattern

GeomPolygonPattern

GeomCrossbarPattern

GeomRibbonPattern

GeomAreaPattern

GeomDensityPattern

GeomPolygonPattern

GeomMapPattern

GeomSfPattern

GeomTilePattern

GeomViolinPattern
}
\keyword{datasets}
