% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-read.R
\name{img_read}
\alias{img_read}
\title{Fetch a given path or URL as a magick image}
\usage{
img_read(filename)
}
\arguments{
\item{filename}{filename or URL}
}
\value{
magick image
}
\description{
Fetch a given path or URL as a magick image
}
