% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-read.R
\name{img_read_as_array}
\alias{img_read_as_array}
\title{Fetch a given path or URL as a 3D RGB array of values}
\usage{
img_read_as_array(
  filename,
  width = NULL,
  height = NULL,
  fill_type = "squish",
  gravity = "Center",
  scale = 1,
  filter = "lanczos"
)
}
\arguments{
\item{filename}{filename or URL}

\item{width, height}{if specified scale the image to these dimensions before
converting to an array}

\item{fill_type}{how to fill the image area. 'none', 'fit',
'squish', 'expand', 'tile'}

\item{gravity}{imagemagick gravity option on how to position an image during
a resize/fill}

\item{scale}{scale image prior to tiling. Only for fill_type == 'tile'}

\item{filter}{filter for scaling. default: lanczos. only for fill_type = 'tile'}
}
\value{
3D array. If the image was pure gray, then it will be promoted to
        be an unsaturated RGB image.
}
\description{
Fetch a given path or URL as a 3D RGB array of values
}
