% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-image.R
\name{img_read_as_array_wrapper}
\alias{img_read_as_array_wrapper}
\title{Read a user specified filename/URL as an image}
\usage{
img_read_as_array_wrapper(width, height, params, legend)
}
\arguments{
\item{width}{image dimensions}

\item{height}{image dimensions}

\item{params}{geom parameters from ggplot}

\item{legend}{logical. TRUE if call comes during legend creation, otherwise FALSE.}
}
\value{
array
}
\description{
Read a user specified filename/URL as an image
}
