% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_df.R
\name{is_polygon_df}
\alias{is_polygon_df}
\title{Test if object is polygon_df or NULL}
\usage{
is_polygon_df(x)
}
\arguments{
\item{x}{object}
}
\value{
TRUE if object is polygon_df or NULL
}
\description{
Test if object is polygon_df or NULL
}
\examples{
  df <- create_polygon_df(x = c(0, 0, 1, 1), y = c(0, 1, 1, 0))
  is_polygon_df(df)
}
