% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick.R
\docType{data}
\name{magick_gravity_names}
\alias{magick_gravity_names}
\title{Gravity types for 'image' positioning}
\format{
An object of class \code{character} of length 13.
}
\usage{
magick_gravity_names
}
\description{
Gravity types for 'image' positioning
}
\keyword{datasets}
