% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick-pattern.R
\docType{data}
\name{magick_pattern_names}
\alias{magick_pattern_names}
\alias{magick_pattern_intensity_names}
\alias{magick_pattern_stripe_names}
\title{Names of patterns available in image magick, plus subsets for shaded intensity and stripes}
\format{
An object of class \code{character} of length 54.

An object of class \code{character} of length 21.

An object of class \code{character} of length 19.
}
\usage{
magick_pattern_names

magick_pattern_intensity_names

magick_pattern_stripe_names
}
\description{
See \url{http://www.imagemagick.org/script/formats.php} for more information.
}
\keyword{datasets}
