% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aab-utils.R
\name{my_abind}
\alias{my_abind}
\title{abind clone for adding a matrix to an array}
\usage{
my_abind(arr, mat)
}
\arguments{
\item{arr, mat}{array and matrix to be stacked}
}
\value{
new array with matrix added as a new plane at the end of the array
}
\description{
A very cut-down version of \code{abind::abind()} that will only stack a
matrix onto an array.
}
