% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern-array-placeholder.R
\docType{data}
\name{placeholder_names}
\alias{placeholder_names}
\title{All placeholder names}
\format{
An object of class \code{character} of length 26.
}
\usage{
placeholder_names
}
\description{
All placeholder names
}
\keyword{datasets}
