% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_df.R
\name{rotate_polygon_df}
\alias{rotate_polygon_df}
\title{Simple 2D rotation of a polygon about the origin}
\usage{
rotate_polygon_df(polygon_df, angle, aspect_ratio)
}
\arguments{
\item{polygon_df}{polygon data.frame}

\item{angle}{angle in degrees}

\item{aspect_ratio}{aspect ratio}
}
\description{
Simple 2D rotation of a polygon about the origin
}
