% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern.R
\name{scale_discrete}
\alias{scale_discrete}
\alias{scale_pattern_continuous}
\alias{scale_pattern_discrete}
\alias{scale_pattern_type_continuous}
\alias{scale_pattern_type_discrete}
\alias{scale_pattern_subtype_continuous}
\alias{scale_pattern_subtype_discrete}
\alias{scale_pattern_filename_continuous}
\alias{scale_pattern_filename_discrete}
\alias{scale_pattern_filter_continuous}
\alias{scale_pattern_filter_discrete}
\alias{scale_pattern_gravity_continuous}
\alias{scale_pattern_gravity_discrete}
\alias{scale_pattern_orientation_continuous}
\alias{scale_pattern_orientation_discrete}
\alias{scale_pattern_grid_continuous}
\alias{scale_pattern_grid_discrete}
\title{Scales for discrete pattern aesthetics}
\usage{
scale_pattern_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = c("stripe", "crosshatch", "circle"),
  trans = "identity",
  guide = "legend"
)

scale_pattern_discrete(
  ...,
  choices = c("stripe", "crosshatch", "circle"),
  guide = "legend"
)

scale_pattern_type_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = NULL,
  trans = "identity",
  guide = "legend"
)

scale_pattern_type_discrete(..., choices = NULL, guide = "legend")

scale_pattern_subtype_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = NULL,
  trans = "identity",
  guide = "legend"
)

scale_pattern_subtype_discrete(..., choices = NULL, guide = "legend")

scale_pattern_filename_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = NULL,
  trans = "identity",
  guide = "legend"
)

scale_pattern_filename_discrete(..., choices = NULL, guide = "legend")

scale_pattern_filter_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = c("lanczos", "box", "spline", "cubic"),
  trans = "identity",
  guide = "legend"
)

scale_pattern_filter_discrete(
  ...,
  choices = c("lanczos", "box", "spline", "cubic"),
  guide = "legend"
)

scale_pattern_gravity_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = c("center", "north", "south", "east", "west", "northeast", "northwest",
    "southeast", "southwest"),
  trans = "identity",
  guide = "legend"
)

scale_pattern_gravity_discrete(
  ...,
  choices = c("center", "north", "south", "east", "west", "northeast", "northwest",
    "southeast", "southwest"),
  guide = "legend"
)

scale_pattern_orientation_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = c("horizontal", "vertical", "radial"),
  trans = "identity",
  guide = "legend"
)

scale_pattern_orientation_discrete(
  ...,
  choices = c("horizontal", "vertical", "radial"),
  guide = "legend"
)

scale_pattern_grid_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = c("square", "hex"),
  trans = "identity",
  guide = "legend"
)

scale_pattern_grid_discrete(
  ...,
  choices = c("square", "hex"),
  guide = "legend"
)
}
\arguments{
\item{name, breaks, labels, limits, trans, guide, ...}{See
\code{{ggplot2}} documentation for more information on scales.}

\item{choices}{vector of values to choose from.}
}
\description{
Scales for discrete pattern aesthetics
}
