% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern-alpha.R
\name{scale_pattern_alpha_continuous}
\alias{scale_pattern_alpha_continuous}
\alias{scale_pattern_alpha_discrete}
\alias{scale_pattern_alpha_ordinal}
\title{Alpha transparency scales}
\usage{
scale_pattern_alpha_continuous(..., range = c(0.1, 1))

scale_pattern_alpha_discrete(...)

scale_pattern_alpha_ordinal(..., range = c(0.1, 1))
}
\arguments{
\item{...}{Other arguments passed on to \code{\link[=continuous_scale]{continuous_scale()}}, \link{binned_scale},
or \code{\link[=discrete_scale]{discrete_scale()}} as appropriate, to control name, limits,
breaks, labels and so forth.}

\item{range}{Output range of alpha values. Must lie between 0 and 1.}
}
\value{
A \link[ggplot2:ggplot2-ggproto]{ggplot2::Scale} object.
}
\description{
See \code{ggplot2::scale_alpha()} for details.
}
