% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern.R
\name{scale_pattern_angle_continuous}
\alias{scale_pattern_angle_continuous}
\alias{scale_pattern_angle_discrete}
\title{Scales for pattern_angle}
\usage{
scale_pattern_angle_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(0, 90),
  trans = "identity",
  guide = "legend"
)

scale_pattern_angle_discrete(..., range = c(0, 90))
}
\arguments{
\item{name, breaks, labels, limits, range, trans, guide, ...}{See
\code{ggplot2} documentation for more information on scales.}
}
\description{
Scales for pattern_angle
}
