% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern-viridis.R, R/zxx.r
\name{scale_pattern_colour_viridis_d}
\alias{scale_pattern_colour_viridis_d}
\alias{scale_pattern_fill_viridis_d}
\alias{scale_pattern_fill2_viridis_d}
\alias{scale_pattern_colour_viridis_c}
\alias{scale_pattern_fill_viridis_c}
\alias{scale_pattern_fill2_viridis_c}
\alias{scale_pattern_colour_ordinal}
\alias{scale_pattern_color_ordinal}
\alias{scale_pattern_fill_ordinal}
\alias{scale_pattern_fill2_ordinal}
\alias{scale_pattern_color_viridis_d}
\alias{scale_pattern_color_viridis_c}
\title{Viridis colour scales from viridisLite}
\usage{
scale_pattern_colour_viridis_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  aesthetics = "pattern_colour"
)

scale_pattern_fill_viridis_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  aesthetics = "pattern_fill"
)

scale_pattern_fill2_viridis_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  aesthetics = "pattern_fill2"
)

scale_pattern_colour_viridis_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "pattern_colour"
)

scale_pattern_fill_viridis_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "pattern_fill"
)

scale_pattern_fill2_viridis_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  option = "D",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "pattern_fill2"
)
}
\arguments{
\item{...}{Other arguments passed on to [discrete_scale()],
[continuous_scale()], or [binned_scale] to control name, limits, breaks,
  labels and so forth.}

\item{begin, end, alpha, direction, option, values, space, na.value, guide}{See \code{ggplot2::scale_colour_viridis_d} for more information}

\item{aesthetics}{Character string or vector of character strings listing the
name(s) of the aesthetic(s) that this scale works with. This can be useful, for
example, to apply colour settings to the `colour` and `fill` aesthetics at the
same time, via `aesthetics = c("colour", "fill")`.}
}
\description{
The `viridis` scales provide colour maps that are perceptually uniform in both
colour and black-and-white. They are also designed to be perceived by viewers
with common forms of colour blindness. See also
<https://bids.github.io/colormap/>.
}
