% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern.R
\name{scale_pattern_continuous}
\alias{scale_pattern_continuous}
\alias{scale_pattern_discrete}
\title{Scales for pattern}
\usage{
scale_pattern_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = c("stripe", "crosshatch", "circle"),
  trans = "identity",
  guide = "legend"
)

scale_pattern_discrete(
  ...,
  choices = c("stripe", "crosshatch", "circle"),
  guide = "legend"
)
}
\arguments{
\item{name, breaks, labels, limits, trans, guide, ...}{See
\code{ggplot2} documentation for more information on scales.}

\item{choices}{vector of values to choose from.}
}
\description{
Scales for pattern
}
