% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern.R
\name{scale_pattern_gravity_continuous}
\alias{scale_pattern_gravity_continuous}
\alias{scale_pattern_gravity_discrete}
\title{Scales for pattern_gravity}
\usage{
scale_pattern_gravity_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  choices = c("center", "north", "south", "east", "west", "northeast", "northwest",
    "southeast", "southwest"),
  trans = "identity",
  guide = "legend"
)

scale_pattern_gravity_discrete(
  ...,
  choices = c("center", "north", "south", "east", "west", "northeast", "northwest",
    "southeast", "southwest"),
  guide = "legend"
)
}
\arguments{
\item{name, breaks, labels, limits, trans, guide, ...}{See
\code{ggplot2} documentation for more information on scales.}

\item{choices}{vector of values to choose from.}
}
\description{
Scales for pattern_gravity
}
