% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern-linetype.R
\name{scale_pattern_linetype}
\alias{scale_pattern_linetype}
\alias{scale_pattern_linetype_continuous}
\alias{scale_pattern_linetype_discrete}
\title{Scale for line patterns}
\usage{
scale_pattern_linetype(..., na.value = "blank")

scale_pattern_linetype_continuous(...)

scale_pattern_linetype_discrete(..., na.value = "blank")
}
\arguments{
\item{...}{see \code{ggplot2::scale_linetype} for more information}

\item{na.value}{The linetype to use for `NA` values.}
}
\description{
Default line types based on a set supplied by Richard Pearson,
University of Manchester. Continuous values can not be mapped to
line types.
}
