% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern.R
\name{scale_pattern_phase_continuous}
\alias{scale_pattern_phase_continuous}
\alias{scale_pattern_phase_discrete}
\title{Scales for pattern_phase}
\usage{
scale_pattern_phase_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = NULL,
  trans = "identity",
  guide = "legend"
)

scale_pattern_phase_discrete(..., range = NULL)
}
\arguments{
\item{name, breaks, labels, limits, range, trans, guide, ...}{See
\code{ggplot2} documentation for more information on scales.}
}
\description{
Scales for pattern_phase
}
