% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern-size.R
\name{scale_pattern_size_continuous}
\alias{scale_pattern_size_continuous}
\alias{scale_pattern_size}
\alias{scale_pattern_size_discrete}
\alias{scale_pattern_size_ordinal}
\title{Scales for area or radius}
\usage{
scale_pattern_size_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  trans = "identity",
  guide = "legend"
)

scale_pattern_size(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(1, 6),
  trans = "identity",
  guide = "legend"
)
}
\arguments{
\item{name, breaks, labels, limits, trans, guide}{See \code{ggplot2::scale_size} for more information}

\item{range}{a numeric vector of length 2 that specifies the minimum and
maximum size of the plotting symbol after transformation.}
}
\value{
A \link[ggplot2:ggplot2-ggproto]{ggplot2::Scale} object.
}
\description{
Scales for area or radius
}
