% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-pattern.R
\name{scale_pattern_xoffset_continuous}
\alias{scale_pattern_xoffset_continuous}
\alias{scale_pattern_xoffset_discrete}
\title{Scales for pattern_xoffset}
\usage{
scale_pattern_xoffset_continuous(
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  range = c(0.01, 0.1),
  trans = "identity",
  guide = "legend"
)

scale_pattern_xoffset_discrete(..., range = c(0.01, 0.1))
}
\arguments{
\item{name, breaks, labels, limits, range, trans, guide, ...}{See
\code{ggplot2} documentation for more information on scales.}
}
\description{
Scales for pattern_xoffset
}
