% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-magick.R
\name{convert_img_to_array}
\alias{convert_img_to_array}
\title{Convert a magick image to an RGBA array.}
\usage{
convert_img_to_array(img, transpose = TRUE)
}
\arguments{
\item{img}{magick image}

\item{transpose}{transpose the array. default TRUE. Image magick data array
is a transponed representation of the data. Always need to transponse
it to get it the right way up in array format}
}
\value{
RGBA array
}
\description{
Convert a magick image to an RGBA array.
}
