% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpattern-deprecated.R
\name{ggpattern-deprecated}
\alias{ggpattern-deprecated}
\alias{convert_polygon_df_to_polygon_grob}
\alias{convert_polygon_df_to_polygon_sf}
\alias{convert_polygon_sf_to_polygon_df}
\title{Deprecated data/functions}
\usage{
convert_polygon_df_to_polygon_grob(
  polygon_df,
  default.units = "npc",
  gp = gpar()
)

convert_polygon_df_to_polygon_sf(polygon_df, buffer_dist = 0)

convert_polygon_sf_to_polygon_df(mp)
}
\arguments{
\item{polygon_df}{A \code{polygon_df} object}

\item{default.units}{See \code{\link[grid:grid.polygon]{grid::polygonGrob()}}.}

\item{gp}{See \code{\link[grid:grid.polygon]{grid::polygonGrob()}}.}

\item{buffer_dist}{See \code{\link[sf:geos_unary]{sf::st_buffer()}}.}

\item{mp}{A \code{sf} package polygon object}
}
\description{
These data/functions are Deprecated in this release of ggpattern,
they will be marked as Defunct and removed in a future version.
}
