% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpattern-deprecated.R
\name{ggpattern-deprecated}
\alias{ggpattern-deprecated}
\title{Deprecated data/functions}
\description{
These data/functions are Deprecated in this release of ggpattern,
they will be marked as Defunct and removed in a future version.
}
